requirements.txt
streamlit>=1.28.0
openai>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
openai
numpy
sentence-transformers
pandas
PyPDF2
python-dotenv
chromadb
